# [resampleFps] add motion compensated interpolation option