--- a/avidemux_plugins/ADM_videoFilters6/resampleFps/confResampleFps.conf
+++ b/avidemux_plugins/ADM_videoFilters6/resampleFps/confResampleFps.conf
@@ -2,5 +2,5 @@ confResampleFps{
 uint32_t:mode
 uint32_t:newFpsDen
 uint32_t:newFpsNum
-bool:blend
+uint32_t:interpolation
 }