--- a/avidemux_plugins/ADM_videoFilters6/resampleFps/confResampleFps.h
+++ b/avidemux_plugins/ADM_videoFilters6/resampleFps/confResampleFps.h
@@ -5,5 +5,5 @@ typedef struct {
 uint32_t mode;
 uint32_t newFpsDen;
 uint32_t newFpsNum;
-bool blend;
+uint32_t interpolation;
 }confResampleFps;