--- a/avidemux_plugins/ADM_videoFilters6/resampleFps/confResampleFps_desc.cpp
+++ b/avidemux_plugins/ADM_videoFilters6/resampleFps/confResampleFps_desc.cpp
@@ -3,6 +3,6 @@ extern const ADM_paramList confResampleFps_param[]={
  {"mode",offsetof(confResampleFps,mode),"uint32_t",ADM_param_uint32_t},
  {"newFpsDen",offsetof(confResampleFps,newFpsDen),"uint32_t",ADM_param_uint32_t},
  {"newFpsNum",offsetof(confResampleFps,newFpsNum),"uint32_t",ADM_param_uint32_t},
- {"blend",offsetof(confResampleFps,blend),"bool",ADM_param_bool},
+ {"interpolation",offsetof(confResampleFps,interpolation),"uint32_t",ADM_param_uint32_t},
 {NULL,0,NULL}
 };