its experimental / far from perfect, it does not handle large motion well...
it is significantly faster than ffmpeg's minterpolate, but quality lacks behind.
however, if there are no large motions (eg. 20fps->60fps, 30fps->60fps) it is better most of the time than the "Blend" interpolation.
so dont try 1..2..5 fps->60fps conversion :)
there is no scene change detection. visible artifacts can occur.
overall, its not bad for a two afternoon job :)

i dont know cmake at all, so it was a little struggle to make CMakeLists.txt work, i hope its okay in the current state.