@szlldm 2021-06-10
Tested on the same 720p@29.97fps sample (upsampled to 60fps), this is 9 times faster than ffmpeg's minterp. Quality difference is not significant (but sometimes noticable if i watch especially for differences).