@eumagga0x2a 2021-06-10
I apologize for the delays.

I use

```
INCLUDE(vf_plugin)

SET(ADM_vf_resampleFps_SRCS ADM_vidResampleFPS.cpp motin.cpp)

IF(DO_COMMON)
    include(admCheckThreads)
ENDIF(DO_COMMON)

ADD_VIDEO_FILTER(ADM_vf_resampleFps ${ADM_vf_resampleFps_SRCS})
INIT_VIDEO_FILTER(ADM_vf_resampleFps)
INSTALL_VIDEO_FILTER(ADM_vf_resampleFps)

IF(DO_COMMON)
    ADM_LINK_THREAD(ADM_vf_resampleFps)
ENDIF(DO_COMMON)
```
as CMakeLists.txt to build the enhanced resampleFps filter. I haven't tried cross-compilation and building on macOS yet (and I cannot test MSVC, obviously).

The preference FEATURES_THREADING_LAVC is not meant to control anything beyond the scope hinted by its name, i.e. it should not be used to control multi-threading in components other than libavcodec-based video decoders. It is also currently always set to 1 (single thread) when hw accelerated decoding is used.

> overall, its not bad for a two afternoon job :)

It is outright amazing. However, the performance with a 654x480 sample and multi-threading hardcoded to 2 threads (my PC has a dual-core CPU) is 10 fps and 8 fps with multi-threading disabled. My feeling is that when it is that slow, it doesn't matter that it is faster than minterpolate. Personally, I would prefer going for maximum quality no matter how slow, else it feels like finding a middle ground between exiting a highway and continuing straight ahead.