@szlldm 2021-06-11
Thanks for the cmake file!

> However, the performance with a 654x480 sample and multi-threading hardcoded to 2 threads (my PC has a dual-core CPU) is 10 fps and 8 fps with multi-threading disabled.

I had around 35 fps at the mentioned 720p (29.97->60 fps) conversion. With the new commit, i can get 50 fps.

> My feeling is that when it is that slow, it doesn't matter that it is faster than minterpolate. 
IMHO 3-4 fps vs 50 fps is a BIG difference.

> Personally, I would prefer going for maximum quality no matter how slow,

My first attempt was, to port/use minterpolate, but it would be very hard. If someone will port that, it could be a 4th option, and the user could decide between a fast and a slow (but better quality) motion interpolation.

>  else it feels like finding a middle ground between exiting a highway and continuing straight ahead.
Continuing the "car metaphore", quality-wise the Blend option is a wheelchair, my "Motin" is a scooter, and minterpolate is a regular car. (Fps-wise its in reverse :) )