@szlldm 2021-06-11
I had improved on quality. Now there is no big difference on the sample i used.
I have also tried larger upsampling, 29.97->180 fps is 20x faster than minterpolate.
So why 180fps? Chaining resampleFps and changeFps (29.97 -> 180 -> 30 or 60)  you can do slow motion effect.
I will try to add a scene change detector to improve further.