@szlldm 2021-06-12
With a simple trick a could improve the quality a large extent. Now its basically on par with minterpolate in quality (compared on a very limited sample set, due to minterpolate's extreme slowness).