@szlldm 2021-06-12
DONE!
I have fixed the motion estimation errors at the edges. Now i cant perceive any differences between minterpolate and motin (tested on a section of the famous Costa Rica 4k video, downsized to 480p before get slowmotioned).