@eumagga0x2a 2021-06-12
Thank you very much. I pulled it, we'll see what breaks. Please keep maintaining the new feature, I cannot do it for you :-)

The quality is indeed better, it is worth a drop of speed to 3 fps with my sample, IMHO.

Of course, all the motion estimation cannot perform miracles, i.e. handle the image of moving bow of a string player rapidly changing strings back and forth. When resampled from almost 1001/30000 to strictly 1001/30000 with ME, you see two bow tips. In the original video, there is no temporal bleding, all clean :-)

> Continuing the "car metaphore"

There was actually no car metaphore involved, purely the problem of policy. An example of a situation where an agreement based on a trade off (without broadening the scope of negotiotions) is much worse than no agreement at all.